error handling

+++ services/web/server/src/simcore_service_webserver/projects/_trash_api.py
@@ -12,7 +12,7 @@ from simcore_service_webserver.director_v2.exceptions import DirectorServiceErro
 
 from ..director_v2 import api as director_v2_api
 from . import projects_api
-from .exceptions import ProjectLockError, StopProjectBeforeTrashError
+from .exceptions import ProjectLockError, ProjectRunningConflictError, ProjectStopError
 from .models import ProjectPatchExtended
 from .settings import get_plugin_settings
 
@@ -49,28 +49,43 @@ async def trash_project(
     user_id: UserID,
     project_id: ProjectID,
     trashed: bool,
+    forced: bool = True,
 ):
     if trashed:
         # stop first
-        try:
-            await projects_api.remove_project_dynamic_services(
-                user_id=user_id,
-                project_uuid=f"{project_id}",
-                app=app,
-                simcore_user_agent=UNDEFINED_DEFAULT_SIMCORE_USER_AGENT_VALUE,
-                notify_users=False,
-            )
 
-            await director_v2_api.delete_pipeline(
-                app, user_id=user_id, project_id=project_id
+        if forced:
+            try:
+                await projects_api.remove_project_dynamic_services(
+                    user_id=user_id,
+                    project_uuid=f"{project_id}",
+                    app=app,
+                    simcore_user_agent=UNDEFINED_DEFAULT_SIMCORE_USER_AGENT_VALUE,
+                    notify_users=False,
+                )
+
+                await director_v2_api.delete_pipeline(
+                    app, user_id=user_id, project_id=project_id
+                )
+            except (DirectorServiceError, ProjectLockError) as exc:
+                raise ProjectStopError(
+                    project_uuid=project_id,
+                    user_id=user_id,
+                    product_name=product_name,
+                    from_err=exc,
+                ) from exc
+        else:
+
+            running = await director_v2_api.is_pipeline_running(
+                app=app, user_id=user_id, project_id=project_id
             )
-        except (DirectorServiceError, ProjectLockError) as exc:
-            raise StopProjectBeforeTrashError(
-                project_uuid=project_id,
-                user_id=user_id,
-                product_name=product_name,
-                from_err=exc,
-            ) from exc
+            # NOTE: must do here as well for dynamic services but needs refactoring!
+            if running:
+                raise ProjectRunningConflictError(
+                    project_uuid=project_id,
+                    user_id=user_id,
+                    product_name=product_name,
+                )
 
     # mark as trash
     await projects_api.patch_project(
